Реализовал макет формы добавления социальных сетей при регистрации волонтёра
